[EXT][SPARK-23514] Use SessionState.newHadoopConf() to propage hadoop configs set in SQLConf. #20679
A few places in spark-sql were using sc.hadoopConfiguration directly. They should be using sessionState.newHadoopConf() to blend in configs that were set through SQLConf.

Also, for better UX, for these configs blended in from SQLConf, we should consider removing the spark.hadoop prefix, so that the settings are recognized whether or not they were specified by the user.